\name{pelicanKnitr-package}
\alias{pelicanKnitr-package}
\alias{pelicanKnitr}
\docType{package}
\title{
\packageTitle{pelicanKnitr}
}
\description{
\packageDescription{pelicanKnitr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pelicanKnitr}
\packageIndices{pelicanKnitr}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{pelicanKnitr}

Maintainer: \packageMaintainer{pelicanKnitr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
